Separate validation into its own class, at some point we might use it in multiple places and add more advanced functionality.
