+++ resources/assets/js/classes/Definition.js
@@ -1,3 +1,5 @@
+import Validation from './Validation';
+
 export default class Definition {
 
 	static typeMap = {
@@ -21,15 +23,6 @@ export default class Definition {
 		nested     : 'array',
 		collection : 'array',
 		group      : 'object'
-	};
-
-	static messages = {
-		'min':        (val) => `This field needs at least ${val} items.`,
-		'max':        (val) => `This field can't have more than ${val} items.`,
-		'min_value':  (val) => `This number must be more than ${val}.`,
-		'max_value':  (val) => `This number must be less than ${val}.`,
-		'min_length': (val) => `This field must be at least ${val} characters long.`,
-		'max_length': (val) => `This field can't be more than ${val} characters long.`
 	};
 
 	static definitions = {};
@@ -185,91 +178,11 @@ export default class Definition {
 	}
 
 	static transformValidationRule(validationRule, { type }) {
-		let tranformedRule = {};
+		let
+			tranformedRule = {},
+			[rule, value] = validationRule.split(':');
 
-		let [rule, value] = validationRule.split(':');
-
-		if(
-			[
-				'min_value', 'max_value',
-				'min_length', 'max_length',
-				'min', 'max', 'length'
-			].indexOf(rule) !== -1
-		) {
-			value = parseFloat(value, 2);
-		}
-
-		switch(rule) {
-			case 'string':
-				tranformedRule = { type: 'string' };
-				break;
-
-			case 'integer':
-				tranformedRule = { type: 'integer' };
-				break;
-
-			case 'in':
-				tranformedRule = {
-					type: 'enum',
-					enum: value.split(',')
-				};
-				break;
-
-			case 'required':
-			case 'present':
-				tranformedRule = {
-					required: true,
-					message: 'This field is required.'
-				};
-				break;
-
-			// These are possible client-side but don't currently
-			// have a corresponding implementation server-side.
-
-			// case 'number':
-			// 	tranformedRule = { type: 'number' };
-			// 	break;
-
-			// case 'boolean':
-			// 	tranformedRule = { type: 'boolean' };
-			// 	break;
-
-			// case 'float':
-			// 	tranformedRule = { type: 'float' };
-			// 	break;
-
-			// case 'array':
-			// 	tranformedRule = { type: 'array' };
-			// 	break;
-
-			// case 'object':
-			// 	tranformedRule = { type: 'object' };
-			// 	break;
-
-			case 'min':
-			case 'min_value':
-			case 'min_length':
-				tranformedRule = { min: value };
-				break;
-
-			case 'max':
-			case 'max_value':
-			case 'max_length':
-				tranformedRule = { max: value };
-				break;
-
-			case 'length':
-				tranformedRule = { len: value };
-				break;
-
-			case 'regex':
-				tranformedRule = { regexp: value };
-				break;
-		}
-
-		if(Definition.messages[rule]) {
-			tranformedRule.message = Definition.messages[rule](value);
-		}
+		tranformedRule = Validation.transform(rule, value);
 
 		// only infer type validation if it's not explicitly defined
 		if(tranformedRule.type === void 0) {
